In "stop" sleep only after sending the first signal to speedup termination
By default the sleep time between sending term/kill signals is 1s
meaning that previously the termination of each process took *at least*
1s which adds up in multi-process applications. There should be no need
to wait until we send out the first signal to the process to stop.

The subtest "process is_running()" did call "stop()" immediately after
"start()" which can not work anymore with this change as the call to
"stop()" terminates the spawned subprocess before it gets a chance to
write its expected output. Other subtests use a different approach. The
most simple solution I could find is to apply the same as the subtest in
before, i.e. a `sleep 1`. I don't like to put sleeps in code but this
isn't really my code and as the design seem to be around absolute
waiting time I don't want to introduce further change in this step.

I mainly tested this together with isotovideo from
https://github.com/os-autoinst/os-autoinst where the above change helped
to reduce the shutdown time of the overall process stack.