Fix sporadic test failures introduced by fb9bc63
As fb9bc63 deleted the internal timeout within the "stop()" method some
tests need to be adjusted to prevent sporadic test failures same as
already done in fb9bc63 for another test. The fix is done using the same
approach, i.e. adding a `sleep 1` to give a chance to the subprocesses
to print some output.