Fix collect status of fast exiting processes
If a sub process is exiting before the parent process calls
`session->register()` we miss the `collect_status` thus the process end
up in `session->orphan()`.
Also we did a lot of work inside the signal handler context. This is
actually bad as well.

This fix change the work of the signalhandler to a minimum, it just
collect the process info and store it inside of an array.
Then on each `is_running()` call, we process this list and call the
corresponding events.