Fix sporadic 01_run.t failure -- "process is still running"
We encountered flaky 01_run.t tests with the error looks like:
```
 #   Failed test 'process is still running'
 #   at t/01_run.t line 235.
 #          got: '0'
 #     expected: '1'
 # Looks like you failed 1 test of 38.

 #   Failed test 'process execute()'
 #   at t/01_run.t line 302.
```
It occurs in one out of ~300 runs. This fix assume, that we actually
send the SIGTERM before the executed script `term_trap.sh` has setup
it's traphandler to ignore such signal.

The fix just listen on STDOUT of the script to get informed, once the
sighandlers are setted and then continue.

A sleep would be ok as well, but it's harder to judge how long is a good
sleep :).