WIP: Handle concrete process and whole process group when stopping
In case the process group could not be created it makes sense to at least
take care of the concrete process. Otherwise, when `stop` is invoked before
`setpgrp(0, 0)` (or `setpgrp(0, 0)` is for some reason not invoked after
all) `stop` would not be able to do anything.

This fix is similar to c4c793031 but for the `stop` function. However, in
this case the limitation has existed for a long time and not just since
ba7bb383. As far as I can tell it hasn't mattered much in production so
far, though.

Related ticket: https://progress.opensuse.org/issues/170209